+++ src/cli/__init__.py
@@ -1036,14 +1036,54 @@ def doctor() -> None:
     show_default=True,
     help='The location to create an OpenVPN configuration file.',
 )
+@click.option(
+    '--force',
+    is_flag=True,
+    help=(
+        'Overwrite any files and destroy conflicting containers from previous '
+        'uses of this command.'
+    ),
+)
 @dcos_docker.command('setup-mac-network')
-def setup_mac_network(configuration_dst: Path) -> None:
+def setup_mac_network(configuration_dst: Path, force: bool) -> None:
     """
     Set up a network to connect to nodes on macOS.
 
     This creates an OpenVPN configuration file and describes how to use it.
     """
-    create_mac_network(configuration_dst=configuration_dst)
+    if force:
+        destroy_mac_network_containers()
+
+    try:
+        with click_spinner.spinner():
+            create_mac_network(configuration_dst=configuration_dst)
+    except docker.errors.APIError as exc:
+        if exc.status_code == 409:
+            message = (
+                'Error: A DC/OS E2E network container is already running. '
+                'Use --force to destroy conflicting containers.'
+            )
+            click.echo(message, err=True)
+            sys.exit(1)
+        raise
+
+    message = (
+        '1. Install an OpenVPN client such as Tunnelblick '
+        '(https://tunnelblick.net/downloads.html) '
+        'or Shimo (https://www.shimovpn.com).'
+        '\n'
+        '2. Run "open {configuration_dst}".'
+        '\n'
+        '3. If your OpenVPN client is Shimo, edit the new "docker-for-mac" '
+        'profile\'s Advanced settings to deselect "Send all traffic over VPN".'
+        '\n'
+        '4. In your OpenVPN client, connect to the new "docker-for-mac" '
+        'profile.'
+        '\n'
+        '5. Run "dcos-docker doctor" to confirm that everything is working.'
+    ).format(configuration_dst=configuration_dst)
+
+    click.echo(message=message)
 
 
 @dcos_docker.command('destroy-mac-network')
@@ -1052,3 +1092,13 @@ def destroy_mac_network() -> None:
     Destroy containers created by "dcos-docker setup-mac-network".
     """
     destroy_mac_network_containers()
+    message = (
+        'The containers used to allow access to Docker for Mac\'s internal '
+        'networks have been removed.'
+        '\n'
+        '\n'
+        'It may be the case that the "docker-for-mac" profile still exists in '
+        'your OpenVPN client.'
+    )
+
+    click.echo(message=message)