+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Next
 
 - Previously not all volumes were destroyed when destroying a cluster from the CLI or with the ``Docker`` backend.
   This has been resolved.
+  To remove dangling volumes from previous versions, use ``docker volume prune``.
 - Backwards incompatible change: ``mount`` parameters to ``Docker.__init`` now take a ``list`` of ``docker.types.Mount``\s.
 
 2018.04.30.2